+++ packages/web3-eth/src/rpc_method_wrappers.ts
@@ -39,14 +39,15 @@ import {
 	Transaction,
 	TransactionCall,
 	TransactionWithLocalWalletIndex,
+	Web3EthExecutionAPI,
 } from 'web3-types';
 import { Web3Context, Web3PromiEvent } from 'web3-core';
 import { ETH_DATA_FORMAT, FormatType, DataFormat, DEFAULT_RETURN_FORMAT, format } from 'web3-utils';
 import { isBlockTag, isBytes, isNullish, isString } from 'web3-validator';
 import { SignatureError, TransactionError, TransactionRevertError } from 'web3-errors';
+import { ethRpcMethods } from 'web3-rpc-methods';
 
 import { decodeSignedTransaction } from './utils/decode_signed_transaction';
-import * as rpcMethods from './rpc_methods';
 import {
 	accountSchema,
 	blockSchema,
@@ -71,7 +72,6 @@ import { trySendTransaction } from './utils/try_send_transaction';
 // eslint-disable-next-line import/no-cycle
 import { waitForTransactionReceipt } from './utils/wait_for_transaction_receipt';
 import { watchTransactionForConfirmations } from './utils/watch_transaction_for_confirmations';
-import { Web3EthExecutionAPI } from './web3_eth_execution_api';
 import { NUMBER_DATA_FORMAT } from './constants';
 
 /**
@@ -85,7 +85,7 @@ import { NUMBER_DATA_FORMAT } from './constants';
  * ```
  */
 export const getProtocolVersion = async (web3Context: Web3Context<EthExecutionAPI>) =>
-	rpcMethods.getProtocolVersion(web3Context.requestManager);
+	ethRpcMethods.getProtocolVersion(web3Context.requestManager);
 
 // TODO Add returnFormat parameter
 /**
@@ -106,7 +106,7 @@ export const getProtocolVersion = async (web3Context: Web3Context<EthExecutionAP
  * ```
  */
 export const isSyncing = async (web3Context: Web3Context<EthExecutionAPI>) =>
-	rpcMethods.getSyncing(web3Context.requestManager);
+	ethRpcMethods.getSyncing(web3Context.requestManager);
 
 // TODO consider adding returnFormat parameter (to format address as bytes)
 /**
@@ -119,7 +119,7 @@ export const isSyncing = async (web3Context: Web3Context<EthExecutionAPI>) =>
  * ```
  */
 export const getCoinbase = async (web3Context: Web3Context<EthExecutionAPI>) =>
-	rpcMethods.getCoinbase(web3Context.requestManager);
+	ethRpcMethods.getCoinbase(web3Context.requestManager);
 
 /**
  * Checks whether the node is mining or not.
@@ -133,7 +133,7 @@ export const getCoinbase = async (web3Context: Web3Context<EthExecutionAPI>) =>
  * ```
  */
 export const isMining = async (web3Context: Web3Context<EthExecutionAPI>) =>
-	rpcMethods.getMining(web3Context.requestManager);
+	ethRpcMethods.getMining(web3Context.requestManager);
 
 /**
  * @param web3Context ({@link Web3Context}) Web3 configuration object that contains things such as the provider, request manager, wallet, etc.
@@ -152,7 +152,7 @@ export async function getHashRate<ReturnFormat extends DataFormat>(
 	web3Context: Web3Context<EthExecutionAPI>,
 	returnFormat: ReturnFormat,
 ) {
-	const response = await rpcMethods.getHashRate(web3Context.requestManager);
+	const response = await ethRpcMethods.getHashRate(web3Context.requestManager);
 
 	return format({ eth: 'uint' }, response as Numbers, returnFormat);
 }
@@ -174,7 +174,7 @@ export async function getGasPrice<ReturnFormat extends DataFormat>(
 	web3Context: Web3Context<EthExecutionAPI>,
 	returnFormat: ReturnFormat,
 ) {
-	const response = await rpcMethods.getGasPrice(web3Context.requestManager);
+	const response = await ethRpcMethods.getGasPrice(web3Context.requestManager);
 
 	return format({ eth: 'uint' }, response as Numbers, returnFormat);
 }
@@ -196,7 +196,7 @@ export async function getBlockNumber<ReturnFormat extends DataFormat>(
 	web3Context: Web3Context<EthExecutionAPI>,
 	returnFormat: ReturnFormat,
 ) {
-	const response = await rpcMethods.getBlockNumber(web3Context.requestManager);
+	const response = await ethRpcMethods.getBlockNumber(web3Context.requestManager);
 
 	return format({ eth: 'uint' }, response as Numbers, returnFormat);
 }
@@ -227,7 +227,7 @@ export async function getBalance<ReturnFormat extends DataFormat>(
 	const blockNumberFormatted = isBlockTag(blockNumber as string)
 		? (blockNumber as BlockTag)
 		: format({ eth: 'uint' }, blockNumber as Numbers, ETH_DATA_FORMAT);
-	const response = await rpcMethods.getBalance(
+	const response = await ethRpcMethods.getBalance(
 		web3Context.requestManager,
 		address,
 		blockNumberFormatted,
@@ -269,7 +269,7 @@ export async function getStorageAt<ReturnFormat extends DataFormat>(
 	const blockNumberFormatted = isBlockTag(blockNumber as string)
 		? (blockNumber as BlockTag)
 		: format({ eth: 'uint' }, blockNumber as Numbers, ETH_DATA_FORMAT);
-	const response = await rpcMethods.getStorageAt(
+	const response = await ethRpcMethods.getStorageAt(
 		web3Context.requestManager,
 		address,
 		storageSlotFormatted,
@@ -308,7 +308,7 @@ export async function getCode<ReturnFormat extends DataFormat>(
 	const blockNumberFormatted = isBlockTag(blockNumber as string)
 		? (blockNumber as BlockTag)
 		: format({ eth: 'uint' }, blockNumber as Numbers, ETH_DATA_FORMAT);
-	const response = await rpcMethods.getCode(
+	const response = await ethRpcMethods.getCode(
 		web3Context.requestManager,
 		address,
 		blockNumberFormatted,
@@ -390,7 +390,7 @@ export async function getBlock<ReturnFormat extends DataFormat>(
 	let response;
 	if (isBytes(block)) {
 		const blockHashFormatted = format({ eth: 'bytes32' }, block, ETH_DATA_FORMAT);
-		response = await rpcMethods.getBlockByHash(
+		response = await ethRpcMethods.getBlockByHash(
 			web3Context.requestManager,
 			blockHashFormatted as HexString,
 			hydrated,
@@ -399,7 +399,7 @@ export async function getBlock<ReturnFormat extends DataFormat>(
 		const blockNumberFormatted = isBlockTag(block as string)
 			? (block as BlockTag)
 			: format({ eth: 'uint' }, block as Numbers, ETH_DATA_FORMAT);
-		response = await rpcMethods.getBlockByNumber(
+		response = await ethRpcMethods.getBlockByNumber(
 			web3Context.requestManager,
 			blockNumberFormatted,
 			hydrated,
@@ -434,7 +434,7 @@ export async function getBlockTransactionCount<ReturnFormat extends DataFormat>(
 	let response;
 	if (isBytes(block)) {
 		const blockHashFormatted = format({ eth: 'bytes32' }, block, ETH_DATA_FORMAT);
-		response = await rpcMethods.getBlockTransactionCountByHash(
+		response = await ethRpcMethods.getBlockTransactionCountByHash(
 			web3Context.requestManager,
 			blockHashFormatted as HexString,
 		);
@@ -442,7 +442,7 @@ export async function getBlockTransactionCount<ReturnFormat extends DataFormat>(
 		const blockNumberFormatted = isBlockTag(block as string)
 			? (block as BlockTag)
 			: format({ eth: 'uint' }, block as Numbers, ETH_DATA_FORMAT);
-		response = await rpcMethods.getBlockTransactionCountByNumber(
+		response = await ethRpcMethods.getBlockTransactionCountByNumber(
 			web3Context.requestManager,
 			blockNumberFormatted,
 		);
@@ -476,7 +476,7 @@ export async function getBlockUncleCount<ReturnFormat extends DataFormat>(
 	let response;
 	if (isBytes(block)) {
 		const blockHashFormatted = format({ eth: 'bytes32' }, block, ETH_DATA_FORMAT);
-		response = await rpcMethods.getUncleCountByBlockHash(
+		response = await ethRpcMethods.getUncleCountByBlockHash(
 			web3Context.requestManager,
 			blockHashFormatted as HexString,
 		);
@@ -484,7 +484,7 @@ export async function getBlockUncleCount<ReturnFormat extends DataFormat>(
 		const blockNumberFormatted = isBlockTag(block as string)
 			? (block as BlockTag)
 			: format({ eth: 'uint' }, block as Numbers, ETH_DATA_FORMAT);
-		response = await rpcMethods.getUncleCountByBlockNumber(
+		response = await ethRpcMethods.getUncleCountByBlockNumber(
 			web3Context.requestManager,
 			blockNumberFormatted,
 		);
@@ -567,7 +567,7 @@ export async function getUncle<ReturnFormat extends DataFormat>(
 	let response;
 	if (isBytes(block)) {
 		const blockHashFormatted = format({ eth: 'bytes32' }, block, ETH_DATA_FORMAT);
-		response = await rpcMethods.getUncleByBlockHashAndIndex(
+		response = await ethRpcMethods.getUncleByBlockHashAndIndex(
 			web3Context.requestManager,
 			blockHashFormatted as HexString,
 			uncleIndexFormatted,
@@ -576,7 +576,7 @@ export async function getUncle<ReturnFormat extends DataFormat>(
 		const blockNumberFormatted = isBlockTag(block as string)
 			? (block as BlockTag)
 			: format({ eth: 'uint' }, block as Numbers, ETH_DATA_FORMAT);
-		response = await rpcMethods.getUncleByBlockNumberAndIndex(
+		response = await ethRpcMethods.getUncleByBlockNumberAndIndex(
 			web3Context.requestManager,
 			blockNumberFormatted,
 			uncleIndexFormatted,
@@ -645,7 +645,7 @@ export async function getTransaction<ReturnFormat extends DataFormat>(
 		transactionHash,
 		DEFAULT_RETURN_FORMAT,
 	);
-	const response = await rpcMethods.getTransactionByHash(
+	const response = await ethRpcMethods.getTransactionByHash(
 		web3Context.requestManager,
 		transactionHashFormatted,
 	);
@@ -742,7 +742,7 @@ export async function getPendingTransactions<ReturnFormat extends DataFormat>(
 	web3Context: Web3Context<EthExecutionAPI>,
 	returnFormat: ReturnFormat,
 ) {
-	const response = await rpcMethods.getPendingTransactions(web3Context.requestManager);
+	const response = await ethRpcMethods.getPendingTransactions(web3Context.requestManager);
 
 	return response.map(transaction =>
 		formatTransaction(transaction as unknown as Transaction, returnFormat),
@@ -811,7 +811,7 @@ export async function getTransactionFromBlock<ReturnFormat extends DataFormat>(
 	let response;
 	if (isBytes(block)) {
 		const blockHashFormatted = format({ eth: 'bytes32' }, block, ETH_DATA_FORMAT);
-		response = await rpcMethods.getTransactionByBlockHashAndIndex(
+		response = await ethRpcMethods.getTransactionByBlockHashAndIndex(
 			web3Context.requestManager,
 			blockHashFormatted as HexString,
 			transactionIndexFormatted,
@@ -820,7 +820,7 @@ export async function getTransactionFromBlock<ReturnFormat extends DataFormat>(
 		const blockNumberFormatted = isBlockTag(block as string)
 			? (block as BlockTag)
 			: format({ eth: 'uint' }, block as Numbers, ETH_DATA_FORMAT);
-		response = await rpcMethods.getTransactionByBlockNumberAndIndex(
+		response = await ethRpcMethods.getTransactionByBlockNumberAndIndex(
 			web3Context.requestManager,
 			blockNumberFormatted,
 			transactionIndexFormatted,
@@ -887,7 +887,7 @@ export async function getTransactionReceipt<ReturnFormat extends DataFormat>(
 		transactionHash,
 		DEFAULT_RETURN_FORMAT,
 	);
-	const response = await rpcMethods.getTransactionReceipt(
+	const response = await ethRpcMethods.getTransactionReceipt(
 		web3Context.requestManager,
 		transactionHashFormatted,
 	);
@@ -929,7 +929,7 @@ export async function getTransactionCount<ReturnFormat extends DataFormat>(
 	const blockNumberFormatted = isBlockTag(blockNumber as string)
 		? (blockNumber as BlockTag)
 		: format({ eth: 'uint' }, blockNumber as Numbers, ETH_DATA_FORMAT);
-	const response = await rpcMethods.getTransactionCount(
+	const response = await ethRpcMethods.getTransactionCount(
 		web3Context.requestManager,
 		address,
 		blockNumberFormatted,
@@ -1113,7 +1113,7 @@ export function sendTransaction<
 							transactionHash = await trySendTransaction(
 								web3Context,
 								async (): Promise<string> =>
-									rpcMethods.sendRawTransaction(
+									ethRpcMethods.sendRawTransaction(
 										web3Context.requestManager,
 										signedTransaction.rawTransaction,
 									),
@@ -1123,7 +1123,7 @@ export function sendTransaction<
 							transactionHash = await trySendTransaction(
 								web3Context,
 								async (): Promise<string> =>
-									rpcMethods.sendTransaction(
+									ethRpcMethods.sendTransaction(
 										web3Context.requestManager,
 										transactionFormatted as Partial<TransactionWithSenderAPI>,
 									),
@@ -1331,7 +1331,7 @@ export function sendSignedTransaction<
 						const transactionHash = await trySendTransaction(
 							web3Context,
 							async (): Promise<string> =>
-								rpcMethods.sendRawTransaction(
+								ethRpcMethods.sendRawTransaction(
 									web3Context.requestManager,
 									signedTransactionFormattedHex,
 								),
@@ -1461,7 +1461,7 @@ export async function sign<ReturnFormat extends DataFormat>(
 		);
 	}
 
-	const response = await rpcMethods.sign(
+	const response = await ethRpcMethods.sign(
 		web3Context.requestManager,
 		addressOrIndex,
 		messageFormatted,
@@ -1526,7 +1526,7 @@ export async function signTransaction<ReturnFormat extends DataFormat>(
 	transaction: Transaction,
 	returnFormat: ReturnFormat,
 ) {
-	const response = await rpcMethods.signTransaction(
+	const response = await ethRpcMethods.signTransaction(
 		web3Context.requestManager,
 		formatTransaction(transaction, ETH_DATA_FORMAT),
 	);
@@ -1566,7 +1566,7 @@ export async function call<ReturnFormat extends DataFormat>(
 		? (blockNumber as BlockTag)
 		: format({ eth: 'uint' }, blockNumber as Numbers, ETH_DATA_FORMAT);
 
-	const response = await rpcMethods.call(
+	const response = await ethRpcMethods.call(
 		web3Context.requestManager,
 		formatTransaction(transaction, ETH_DATA_FORMAT),
 		blockNumberFormatted,
@@ -1615,7 +1615,7 @@ export async function estimateGas<ReturnFormat extends DataFormat>(
 		? (blockNumber as BlockTag)
 		: format({ eth: 'uint' }, blockNumber as Numbers, ETH_DATA_FORMAT);
 
-	const response = await rpcMethods.estimateGas(
+	const response = await ethRpcMethods.estimateGas(
 		web3Context.requestManager,
 		transactionFormatted,
 		blockNumberFormatted,
@@ -1675,7 +1675,7 @@ export async function getLogs<ReturnFormat extends DataFormat>(
 	filter: Filter,
 	returnFormat: ReturnFormat,
 ) {
-	const response = await rpcMethods.getLogs(web3Context.requestManager, filter);
+	const response = await ethRpcMethods.getLogs(web3Context.requestManager, filter);
 
 	const result = response.map(res => {
 		if (typeof res === 'string') {
@@ -1705,7 +1705,7 @@ export async function getChainId<ReturnFormat extends DataFormat>(
 	web3Context: Web3Context<EthExecutionAPI>,
 	returnFormat: ReturnFormat,
 ) {
-	const response = await rpcMethods.getChainId(web3Context.requestManager);
+	const response = await ethRpcMethods.getChainId(web3Context.requestManager);
 
 	return format(
 		{ eth: 'uint' },
@@ -1803,7 +1803,7 @@ export async function getProof<ReturnFormat extends DataFormat>(
 		? (blockNumber as BlockTag)
 		: format({ eth: 'uint' }, blockNumber as Numbers, ETH_DATA_FORMAT);
 
-	const response = await rpcMethods.getProof(
+	const response = await ethRpcMethods.getProof(
 		web3Context.requestManager,
 		address,
 		storageKeysFormatted,
@@ -1899,7 +1899,7 @@ export async function getFeeHistory<ReturnFormat extends DataFormat>(
 		NUMBER_DATA_FORMAT,
 	);
 
-	const response = await rpcMethods.getFeeHistory(
+	const response = await ethRpcMethods.getFeeHistory(
 		web3Context.requestManager,
 		blockCountFormatted,
 		newestBlockFormatted,